bug(dates): render dates - fix render date bug [Finishies #159398942]
